Docs for imposition impls.
