--- ratpack-core/src/main/java/ratpack/impose/ForcePortImposition.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ratpack.impose;
-
-public final class ForcePortImposition implements Imposition {
-
-  private final int port;
-
-  private ForcePortImposition(int port) {
-    this.port = port;
-  }
-
-  public static ForcePortImposition ephemeral() {
-    return of(0);
-  }
-
-  public static ForcePortImposition of(int port) {
-    return new ForcePortImposition(port);
-  }
-
-  public int getPort() {
-    return port;
-  }
-
-}
