+++ ratpack-groovy/src/test/groovy/ratpack/groovy/ScriptBackedServer.groovy
@@ -16,7 +16,7 @@
 
 package ratpack.groovy
 
-import ratpack.impose.ForcePortImposition
+import ratpack.impose.ForceServerListenPortImposition
 import ratpack.impose.Impositions
 import ratpack.server.RatpackServer
 import ratpack.server.StartupFailureException
@@ -35,7 +35,7 @@ class ScriptBackedServer implements RatpackServer {
 
   @Override
   void start() throws StartupFailureException {
-    nestedServer = ServerCapturer.capture(Impositions.of { it.add(ForcePortImposition.ephemeral()) }, { -> starter.run() })
+    nestedServer = ServerCapturer.capture(Impositions.of { it.add(ForceServerListenPortImposition.ephemeral()) }, { -> starter.run() })
 
     def stopAt = System.currentTimeMillis() + 10000
     while (System.currentTimeMillis() < stopAt && (nestedServer == null || !nestedServer.running)) {
