+++ ratpack-manual/src/test/groovy/ratpack/manual/snippets/fixture/ServerCaptureSnippetExecuter.java
@@ -16,7 +16,7 @@
 
 package ratpack.manual.snippets.fixture;
 
-import ratpack.impose.ForcePortImposition;
+import ratpack.impose.ForceServerListenPortImposition;
 import ratpack.impose.Impositions;
 import ratpack.manual.snippets.TestCodeSnippet;
 import ratpack.manual.snippets.executer.SnippetExecuter;
@@ -39,7 +39,7 @@ public abstract class ServerCaptureSnippetExecuter implements SnippetExecuter {
   @Override
   public void execute(TestCodeSnippet snippet) throws Exception {
     withServer(ServerCapturer.capture(
-      Impositions.of(i -> i.add(ForcePortImposition.ephemeral())), () -> executer.execute(snippet))
+      Impositions.of(i -> i.add(ForceServerListenPortImposition.ephemeral())), () -> executer.execute(snippet))
     );
   }
